move test to ids_test.go
